+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 
-## vx.x.x - xxxx-xx-xx
+## v2.3.0 - xxxx-xx-xx
+
+### New Features:
+
+* On-foot murders are now tracked and are independent from ship murders.
 
 ### Changes:
 
@@ -31,6 +35,7 @@
 * Moved the overlay text - both tick time and tick alerts - a little to the left to allow for differences in text placement between machines.
 * When using new modern Discord format, don't create empty posts and delete previous post if content is empty.
 * Minor change to 'CMDRs' button image to make it clearer in dark mode.
+* A limit of 60 is now applied to the number of tabs shown in the activity window, as too many tabs could crash the plugin.
 
 
 ## v2.1.0 - 2022-12-05